+++ src/compounds/product-table/src/components/cell-cta.tsx
@@ -19,9 +19,9 @@ const ProductTableCellCta: React.FC<CellCtaProps> = ({
   const { isCard } = React.useContext(CardContext)
   const forcedMobile = forceMobile(isCard)
   const buttonWrapperStyling = {
-    flex: forcedMobile([1, undefined, 'initial']),
+    flex: forcedMobile([1, null, 'initial']),
     width: forcedMobile(['auto', '100%']),
-    display: forcedMobile(['flex', undefined, 'block']),
+    display: forcedMobile(['flex', null, 'block']),
     variant: headerImage
       ? 'compounds.product-table.variants.redesign.cellCta.buttonWrapper'
       : 'compounds.product-table.cellCta.buttonWrapper'
@@ -32,7 +32,8 @@ const ProductTableCellCta: React.FC<CellCtaProps> = ({
         display: 'flex',
         alignItems: 'center',
         justifyContent: 'center',
-        flexDirection: forcedMobile(['column', 'row', 'column'])
+        flexDirection: forcedMobile(['column', 'row', 'column']),
+        variant: 'compounds.product-table.cta'
       }}
     >
       <div sx={{ ...buttonWrapperStyling }}>{primary}</div>