+++ src/compounds/interactive-tabs/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.3](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.interactive-tabs@2.1.2...@uswitch/trustyle.interactive-tabs@2.1.3) (2020-12-02)
+
+**Note:** Version bump only for package @uswitch/trustyle.interactive-tabs
+
+
+
+
+
+## [2.1.2](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.interactive-tabs@2.1.1...@uswitch/trustyle.interactive-tabs@2.1.2) (2020-11-30)
+
+**Note:** Version bump only for package @uswitch/trustyle.interactive-tabs
+
+
+
+
+
+
 ## [2.1.1](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.interactive-tabs@2.1.0...@uswitch/trustyle.interactive-tabs@2.1.1) (2020-11-16)
 
 **Note:** Version bump only for package @uswitch/trustyle.interactive-tabs