+++ src/compounds/sponsored-product/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.6](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product@3.5.5...@uswitch/trustyle.sponsored-product@3.5.6) (2020-12-10)
+
+**Note:** Version bump only for package @uswitch/trustyle.sponsored-product
+
+
+
+
+
+## [3.5.5](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product@3.5.4...@uswitch/trustyle.sponsored-product@3.5.5) (2020-12-10)
+
+**Note:** Version bump only for package @uswitch/trustyle.sponsored-product
+
+
+
+
+
+
+## [3.5.4](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product@3.5.3...@uswitch/trustyle.sponsored-product@3.5.4) (2020-12-02)
+
+**Note:** Version bump only for package @uswitch/trustyle.sponsored-product
+
+
+
+
+
+## [3.5.3](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product@3.5.2...@uswitch/trustyle.sponsored-product@3.5.3) (2020-11-30)
+
+**Note:** Version bump only for package @uswitch/trustyle.sponsored-product
+
+
+
+
+
+
+## [3.5.2](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product@3.5.1...@uswitch/trustyle.sponsored-product@3.5.2) (2020-11-25)
+
+**Note:** Version bump only for package @uswitch/trustyle.sponsored-product
+
+
+
+
+
 ## [3.5.1](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.sponsored-product@3.5.0...@uswitch/trustyle.sponsored-product@3.5.1) (2020-11-24)
 
 **Note:** Version bump only for package @uswitch/trustyle.sponsored-product