+++ src/elements/bullet-list-highlight/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.3](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.bullet-list-highlight@0.4.2...@uswitch/trustyle.bullet-list-highlight@0.4.3) (2020-12-02)
+
+**Note:** Version bump only for package @uswitch/trustyle.bullet-list-highlight
+
+
+
+
+
+## [0.4.2](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.bullet-list-highlight@0.4.1...@uswitch/trustyle.bullet-list-highlight@0.4.2) (2020-11-30)
+
+**Note:** Version bump only for package @uswitch/trustyle.bullet-list-highlight
+
+
+
+
+
+
 ## [0.4.1](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.bullet-list-highlight@0.4.0...@uswitch/trustyle.bullet-list-highlight@0.4.1) (2020-11-16)
 
 **Note:** Version bump only for package @uswitch/trustyle.bullet-list-highlight